+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/tasks/aware/TestableAware.kt
@@ -2,8 +2,13 @@
 
 package org.jetbrains.intellij.platform.gradle.tasks.aware
 
+import org.gradle.api.file.ConfigurableFileCollection
 import org.gradle.api.file.DirectoryProperty
+import org.gradle.api.tasks.InputFiles
 import org.gradle.api.tasks.Internal
+import org.gradle.api.tasks.PathSensitive
+import org.gradle.api.tasks.PathSensitivity
+import org.jetbrains.intellij.platform.gradle.Constants.Configurations
 import org.jetbrains.intellij.platform.gradle.extensions.IntelliJPlatformExtension
 import org.jetbrains.intellij.platform.gradle.tasks.PrepareTestTask
 import org.jetbrains.intellij.platform.gradle.tasks.TestIdeTask
@@ -13,6 +18,15 @@ import org.jetbrains.intellij.platform.gradle.tasks.TestIdeTask
  * used for configuring `test` and keeping it immutable.
  */
 interface TestableAware : CoroutinesJavaAgentAware, RuntimeAware, SandboxAware {
+
+    /**
+     * Holds the [Configurations.INTELLIJ_PLATFORM_TEST_CLASSPATH] configuration, which by default contains
+     * [Configurations.INTELLIJ_PLATFORM_DEPENDENCY], [Configurations.INTELLIJ_PLATFORM_DEPENDENCIES],
+     * and [Configurations.INTELLIJ_PLATFORM_TEST_DEPENDENCIES] configurations.
+     */
+    @get:InputFiles
+    @get:PathSensitive(PathSensitivity.RELATIVE)
+    val intellijPlatformTestClasspathConfiguration: ConfigurableFileCollection
 
     /**
      * Specifies the directory where the plugin artifacts are to be placed.